secret_guide/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
django==4.2.2
django-bootstrap5==23.3
django-cleanup==8.0.0
django-environ==0.10.0
gevent==22.10.2
greenlet==2.0.2 ; platform_python_implementation == 'CPython'
gunicorn==20.1.0
idna==3.4 ; python_version >= '3.5'
pillow==9.5.0
requests==2.31.0
setuptools==68.0.0 ; python_version >= '3.7'
sqlparse==0.4.4 ; python_version >= '3.5'
urllib3==2.0.3 ; python_version >= '3.7'
zope.event==4.6
zope.interface==6.0 ; python_version >= '3.7'
